Implement frame-pointer stack walker and profiling test infrastructure.
Replaced the stub emb_stack_walk with a real implementation that reads thread state (arm64 + x86_64), walks the frame pointer chain with ptrauth stripping, and uses stack bounds to guard against unsafe memory access.

The frame type is now uintptr_t rather than uint64_t for better C spec adherence (this is mostly cosmetic since all pointers on Apple hardware is 64 bits now).

The tests can be run with and without frame pointers (EmbraceProfilingTestSupportNoFP target), but these won't pass until KSCrash is updated.

Added sampler benchmark tests which will also measure no-frame-pointer fallback mode once available.

+++ Sources/EmbraceProfiling/ProfilingEngine.swift
@@ -125,11 +125,11 @@ public final class ProfilingEngine {
             return samples
         }
 
-        private func fakeSampleFrames(seed: UInt64) -> [UInt64] {
+        private func fakeSampleFrames(seed: UInt64) -> [UInt] {
             // Simulate main thread call stacks with realistic arm64 addresses.
 
             // Usually the base will look like this:
-            let runLoopBase: [UInt64] = [
+            let runLoopBase: [UInt] = [
                 0x00000001_8B6E6000,  // CoreFoundation: CFRunLoopRunSpecific
                 0x00000001_8B6E5800,  // CoreFoundation: __CFRunLoopRun
                 0x00000001_8B6E2400,  // CoreFoundation: __CFRunLoopDoSources0
@@ -144,7 +144,7 @@ public final class ProfilingEngine {
             ]
 
             // Varying stack tops
-            var top: [UInt64]
+            var top: [UInt]
             switch seed % 5 {
             case 0:
                 top = [
